meta-ibm: p10bmc: add a MCTP DISTRO_FEATURES switch
The switch enables AF_MCTP and adds the `mctp` tool and `mctpd` so we
can setup and manage an in-kernel MCTP network.

Enable it for p10bmc and also add CONFIG_TUN so we can use a tun device
to transport packets between the kernel and the astlpc interface
implemented in libmctp.

Change-Id: I21d0af9e751f91975c760b4b4708b3b401b95417
Signed-off-by: Rashmica Gupta <rashmica@linux.ibm.com>