meta-phosphor: add phosphor-static-norootfs-init recipe
There are two existing recipes for early mounting procedures:
  * preinit-mounts - for obmc-ubi-fs
  * phosphor-mmc-init - for phosphor-mmc

Add one for the static-norootfs image type.  This one is similar to
the preinit-mounts, except for it mounts /var as an overlay to avoid
violating FHS expectations (See openbmc/openbmc#3658 as an example of
why this is problematic).  I did not directly change preinit-mounts and
fix it because there are likely field upgrade issues in doing this for
existing users.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I62ab288c2d139cfecadeb191d576e2753120e123