Add dev packages to openbmc yocto sdk
In the current state, the master sdk does not contain the
following packages, hence it is currently broken.
 - nlohmann-json
 - function2
 - cli11
 - libcereal

Most of the the OpenBMC repositories depend on the above mentioned
packages and does not build in their absence. This commit would
attempt to fix that.

Tested with a newly created SDK with this change and observed that
the header files of above packages are placed in the sdk.

Change-Id: I3b8dde8f5242279f856f3d3ecd3b78a1e93f6ffd
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>