meta-phosphor: add offset for uboot environment
The previous static-image building code only confirmed that the
u-boot image was smaller than the offset for the kernel image, but
most systems place the u-boot environment after the u-boot image.
It is possible that we have a u-boot image bigger than the u-boot
partition, which is then overwritten on boot by the u-boot env.
Check for this at image build time and prevent it from happening.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I3a1e6654272cedb3e840bbea4771ae4e464c4ecd