meta-phosphor: static-norootfs: add manifest JSON to image
It is desirable to have a single image for both manufacturing and
updates instead of the separate MTD and tarball.  Define a JSON
format and populate it, based on a combination of the current MANIFEST
data and learnings from the Meta verified-boot metadata format[1,2].
This JSON file is then embedded into the image itself, so that update
tools can determine the compatibility and correctness of the image
before applying.

An example manifest JSON can be found at [3].

[1]: https://github.com/facebook/openbmc/blob/eda9e95e2d94feaa2749f2a8e97e237bfe7b38e3/tools/flashy/lib/validate/partition/p_fbmeta.go
[2]: https://github.com/facebook/openbmc/blob/eda9e95e2d94feaa2749f2a8e97e237bfe7b38e3/meta-facebook/classes/fbobmc-image.bbclass#L200
[3]: https://gist.github.com/williamspatrick/9bd5e39909df2a3cad4894dc926d5c1c

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Iff0b7a37194297d49ca4f3d6077d5995452d7c72