meta-ibm: Add new board SBP1
Add support for the IBM SBP1 board.

It's an evaluation platform based on Aspeed AST2600 SoC powering
a 4 socket Intel Sapphire Rapids platform.

The main difference to regular platforms is the non-existence of a
CPLD/FPGA. Everything on this board is directly connected to and
controlled by BMC including power sequence for PCH/CPU.

The power-sequencing makes use of in kernel PMBUS drivers that are
currently upstreamed and needs to be backported to OpenBMC kernel fork.

Change-Id: I34a0b70e1f7f6c1ab2be38e8d66d5912950e0d82
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Naresh Solanki <Naresh.Solanki@9elements.com>
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>