meta-fii: meta-mori: Add BIOS accessing library functions
The BMC accesses the BIOS SPI during flashing, but the BMC will
potentially need to access it while the host is running. Add library
functions that performs a handshake to request access to it. Update
current BIOS SPI access to use the functions. Also fix shellcheck
error.

Tested: Flashed BMC image on unit. Ensured that BIOS flash is still
successful.

Change-Id: Icc91802647a6967694fca3e57bbc56f2c1db9189
Signed-off-by: Kyle Nieman <kyle.nieman@fii-na.com>