meta-ibm:witherspoon: no installation in templated targets
Upstream yocto introduced a change via e510222 (systemd-systemctl:
fix instance template WantedBy symlink construction).

This fixes a bug that we in OpenBMC had been taking advantage of in that
we were able to document our templated target dependencies without it
actually doing anything. The real installation of services within
targets occurs in our bitbake recipes due to the complexity of chassis
and host instances on a per machine basis.

Leave the dependency information in the service files but comment them
out. It's useful to be able to look at a service and understand which
targets it's going to be installed into by the bitbake recipes.

Change-Id: I1bf3bc6ce7a79d6c5847f5f1b4e6a97ca67332f8
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>