phosphor-software-manager: mmc: Add fsck.ext*
Add the fsck.ext* binaries to the mmc images so that the software
updater can run fsck on the alternate image on every boot to catch
corruption in case the BMC rebooted in the middle of a code update for
example.

The dependency needs to be added in the RDEPENDS section instead of the
PACKAGECONFIG because the packages created by this recipe are named
${PN}-* and adding a dependency via the PACKAGECONFIG adds the RDEPENDS
to ${PN} which is not a valid package.

Tested: Verified that the fsck.ext* binaries are available in /sbin/.

Change-Id: Iee56d489c048a712b17f8befd9ea6c01b6625fbe
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>