meta-google: gbmc-mac-config: Fix setting MAC address
The gbmc-mac-config service is used to change the address to match the
value stored in the FRU by creating a config file for the interface.
However in commit 93433c165b38b1bbce6a679e43cf0188311551be of
phosphor-networkd, the default files for the interfaces are
consolidated into a single file for creating them. As a result, the
config generated by the service for eth1 isn't applied, and the
randomly generated MAC is used to obtain an IP address. Eventually, all
IP addresses are allocated to the randomly generated MAC addresses, and
the unit is unable to get a valid IP address. The MAC address is now
set using a different method.

Tested: Offline flashed a unit without the changes. Checked that MAC
address didn't match what is stored in the FRU. Flashed with the changes
and verified the MAC address matched the FRU.

Change-Id: Iaca14e86c93b5392def60d666b8b1aa6b03fd716
Signed-off-by: Kyle Nieman <kyle.nieman@fii-na.com>