meta-phosphor: busybox: Enable truncate applet
Truncate allows for efficient creation of fixed-sized "empty" files
among other things. This is an alternative to `dd if=/dev/zero ...`,
which requires over-specification through `bs=` and `count=` parameters.
The bs= parameter in particular can cause significant inefficiencies if
not applied with care. The presence of `truncate` allows us to avoid
these problems.

Change-Id: I495886293bc1bdd0e461cc995a21ba9aef53f4f8
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>