meta-phosphor: simplify FLASH_SIZE specification
FLASH_SIZE was previously defaulted in a few bbclasses and bbappends,
but this can result in it not being defined in all recipes if the
machine has not overridden the FLASH_SIZE.  openbmc/openbmc#3876
reported one such case where the u-boot-fw-utils recipe does not
receive the correct FLASH_SIZE and ends up erasing the wrong flash
region as if it were the u-boot enviroment.  Specify it in the common
include so it is available everywhere.

Resolves openbmc/openbmc#3876.

Change-Id: Iacdde1a5d0081b54321110d655da18bb3846cd14
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>