meta-nuvoton: trusted-firmware-a: change tf-a load address
For compatible with the newest IGPS 3.9.1 design for new memory map.
We need to modify each image load address individually.

The load address of tf-a from 0x00001000 to 0x02000000
The load address of optee-os from 0x00100000 to 0x02100000
The load address of uboot from 0x03008000 to 0x06208000

We've updated tf-a to 2.8 that keep up with meta-arm.
Thus we remove 2.6.0.bb recipe and replace with 2.8.0.bbappend.

Tested: build pass and boot successfully.

Signed-off-by: Tim Lee <timlee660101@gmail.com>
Change-Id: I730f0a677103b89b78897f0d1860a47933c5b7cc