meta-phosphor: remove obmc-rng feature
Since Linux 5.6, the entropy generation provided by rng-tools is now
in the kernel directly.  Yocto has removed all dependencies on rng-tools
for packages that use to use it, such as ssh[1].  There isn't any reason
for us to explicitly include rng-tools any longer.  Remove the feature
and package group that we had that included it.

[1]: https://lore.kernel.org/openembedded-core/20221209070538.72901-1-xiangyu.chen@eng.windriver.com/

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I85cbe7f30bb41045b7fd1170391a083db9e378b7