meta-phosphor: phosphor-ipmi-fru: remove unused native recipes
phosphor-ipmi-fru has some complex YAML merging recipes that are not
even used.  All of the YAML files provided by these recipes are empty
and there is not a single example of a bbappend replacing them with
a non-empty version.  Delete these recipes and some of the complexity
in the merge recipe.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Ie46e1d2c9de86b0b214c521e49d4a5a6365d9b27