meta-google: Remove python3-core from nftables
python3-core is taking a lot of space causing the image to be too large
in some systems. Remove the python package config to remove the
dependencies. This is not done in bbappend because it didn't work when
testing it out.

Tested:
Built with gbs and it doesn't have python3-core when tested with
```
INHERIT += "buildhistory"
BUILDHISTORY_COMMIT = "1"
```
to get the package sizes.

Change-Id: I5040cfb46ad618d622855891a462d53a887433fd
Signed-off-by: Willy Tu <wltu@google.com>