meta-openpower: Reset dump attribute
Reset dump attribute on obmc-host-quiesce target

The pvm_sys_dump_active bios attribute prevents new dump creation when a
dump is already in progress/ is being offloaded. This attribute is
cleared by the PHYP when the dump is offloaded and deleted.

In case of a PHYP TI with host getting into quiesced state, the
attribute is not cleared which prevents further dump creation. To handle
this scenario, the attribute is reset when the host reaches quiesce
state. Adding the op-clear-sys-dump-active service to the
obmc-host-quiesce target cleared the attribute.

Tested:
When the host is quiesced, initiated a user triggered system dump and
tested. The new system dump is created successfully.

Change-Id: I63b3e3c635d7978aec8ca8f8944f34fe28dfddee
Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>