meta-ibm:wspoon: Use gpioset in power-supply-sync
The phosphor-gpio-util that was being used was a custom tool that can be
replaced by gpioset.  It was also just removed from its repo.

The service now has to be a oneshot because otherwise it isn't allowed
to have multiple ExecStart lines.

Tested:
Powered on multiple times, ensured the power supply input history was
cleared on each power on.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Ib181c67b70e2feb109c1a2aabba8d046e7dfe380