meta-facebook: remove IPMI hw from multihost systems
All meta-facebook systems use ipmb to interact with the OCP
debug card.  The multi-host machines also use ipmb to talk with
the BIC, but the 'ipmi-ipmb' package is not considered an 'ipmi-hw'
package (because it is often installed as an addition).  The single
host machine(s) use 'ipmi-kcs'.

Rearrange the inc file specifications so that the multi-host systems
do not get the 'ipmi-kcs' package.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I756984d32d424dd0b08a44b434318ccee6672889