meta-google: gbmc-bridge: Deprecate hardcoded addresses
These addresses should not be used by default for traffic on the
interface, as we may have multiple BMCs using the same hardcoded address
WRT legacy applications. We want multiple BMCs of the same type to be
able to coexist if they have unique addresses and right now they are not
correctly preferring these addresses.

Change-Id: I0d02cbf63b54f2464ae889ea773e032da0e9b650
Signed-off-by: William A. Kennington III <wak@google.com>