meta-ibm: Remove PS input_history associations
On Rainier and Everest the code to put the power supply input history
maximum and average values on D-Bus is being removed and replaced with a
single peak input power sensor.

Remove the associations to them since they are no longer around.

Note: The associations for the new sensor will be handled by the
psu-monitor application.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I07bb6523f60be0e56db5d85932da9193719c7406