fit-image.bbclass: support vendored dts files
The 6.5 kernel moved dts files to sub-directories for many systems.
Yocto already handles this and uses `KERNEL_DTBVENDORED` as a flag to
indicate if the vendored path should be preserved or not.  By default,
the non-vendored paths are use (resulting in a `basename` call to strip
off the vendor segment in many cases).  Add support in the
fit-image.bbclass for the same `KERNEL_DTBVENDORED` flag as an indicator
if the truncated path should be used or not.

The OpenBMC test bump for 6.5 can be found here:
https://gerrit.openbmc.org/c/openbmc/openbmc/+/65884

Tested:
- Confirmed yosemite4 machine builds with 6.5 kernel

Change-Id: I95d20388b65b9a54396ee8a6ead963a94956db25
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>