webui: add new IMAGE_FEATURE
Lots of machines bring the webui in via an RDEPENDS on something
like the "extras" package. The webui for most OpenBMC machines is a
standard feature (as seen in the amount of .bbappend files used to bring
it in) so lets make it an official image feature.

Do not include this new image feature by default in all OpenBMC images
as building the package is non-standard in how it builds. It directly
calls NPM to install required packages so it can not be built offline
and it complicates the software bill of materials.

Tested:
- Confirmed that system1 has webui by default with this when including
  the new IMAGE_FEATURE

Change-Id: Icc9fd5d1025f7c385b0aa0f6a9dd74e3a7354c7b
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>