initfs: Adapt shutdown unmount to systemd v254
Adjust the awk match statement to unmount the ro and rw filesystems
possibly from multiple places.

Systemd version 254 changed the rules for mount propagation and what
is left mounted under run during systemd-shutdown before transfering
to the initramfs shutdown, and now the read-only and read-write mount
points we used to form the overlay are exposed in multiple places.

To avoid matching the optins, restrict the match criteria to field
two (the mountpoint) and anchor to the beginning or end of the field.
Since busybox awk doesn't handle some regular expressions according
to POSIX, just use . to match any single character instead of
non-standard and POSIX undefined escaping of / in the match criteria.

Fixes: openbmc/openbmc#3947
Change-Id: I04e83a7eb0fc2954b3108115a02cd0a73ce67a90
Signed-off-by: Milton Miller <mdmii@outlook.com>