meta-ibm: Add libmctp to witherspoon-tacoma rootfs
Usage of libmctp is deprecated, and ideally, we would want
to be using the in-kernel implementation of mctp (more so as
the upstream pldm repo already supports it).

But, IBM would probably want some more time before it consumes
the in-kernel mctp implementation as the majority of the apps in p10bmc
still relies on the libmctp API's to communicate with host. This would
be a patch to just fix the broken image temporarily until we migrate
to the in-kernel implementation.

Change-Id: I6aaf52cd8ce83cd2263af161d94213ab80fe383a
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>