meta-bletchley: update host state monitoring mechanism
The original checking mechanism only checks whether the current state
is different from the previous state. It does not consider the state
changes during the state check count, so user will see that the final
state is different from the state at the time of the check count.

Add a new array for storing temp host state so that we can notice the
state changes during checking and reset the check count. Also incerease
the check count to 8 to avoid host state changes frequently.

Signed-off-by: Potin Lai <potin.lai@quantatw.com>
Change-Id: I4713b16becd9787d09f1575fa8df91bbb80e9c5c