meta-facebook: remove yosemite4 power control redundant dependencies
Remove two depnedencies being redundant in yosemite4:
1. slot host off was waiting the invalid ipmi host service,
and we won't use Ipmi.Internal.SoftPowerOff.service.

2. obmc-host-shutdown@.target required by obmc-host-reboot:
power cycle and off will be both triggered with this dependency.

For these reasons, remove dependencies,
and it was verified that power control will be fine with changes.

Tested:
host power on/off/cycle: pass

Change-Id: Ieb0e82b5297d05ec552ccec967d9dc41de4aa756
Signed-off-by: Delphine CC Chiu <Delphine_CC_Chiu@wiwynn.com>