Make upstream ipmitool.bb handle most configuration tasks
The openembedded ipmitool bitbake file is being updated to be
responsible for handling more build features for the OpenBMC ipmitool.

The openembedded bb file has already been modified to pull ipmitool
source from the codeberg.org repo. It is preferred that the OpenBMC
ipmitool follow the openembedded recipe for the source repo, and
revision.

The codeberg ipmitool repository includes changes that manage the
IANA numbers file. Instead of using rules to manage IANA numbers in
this layer, use the meta-oe rules. The meta-oe rules assume the IANA
numbers file has already been downloaded locally, improving build
repeatability.

Tested:
Confirmed the ipmitool version is no longer 1.8.18
The --enable/disable-registry-download works as expected.
The iana-enterprise-numbers is copied to
/usr/share/misc/enterprise-numbers by the upstream recipe.

Change-Id: I78231ca857581c92c1c5106f3f32256b5fa88332
Signed-off-by: Johnathan Mantey <johnathanx.mantey@intel.com>