meta-asrock: Initial skeleton of e3c256d4i platform support
This is a single-socket mini-ITX Xeon board with an AST2500 BMC, roughly
an updated version of the e3c246d4i.  Note though that because this
platform uses eSPI instead of LPC, it currently requires some driver
support that's not yet in the OpenBMC kernel to actually boot the host.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: Ib3a13e0ac9f4130c7965ce63f804eb169262fcd3