meta-asrock: e3c256d4i: Add obmc-console config
While on all other currently supported ASRock platforms the Aspeed VUART
provides the host serial console, we haven't had any success using it
over eSPI on e3c256d4i, so instead we're using a pair of SUARTs
connected via the UART mux.

This unforunately requires leaving the AST2500's built-in Super-IO
enabled, because the host firmware needs to access it shortly after
power-on in order to configure its UART output, so we also have to open
up some of the u-boot backdoor config options too.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: Ibdab607d504cebd956a7e23191c09cd0d5c44e50