meta-google: gbmc-ncsi-config: Remove dhcp-relay from small images
We provide a new variable called `GBMC_NCSI_DHCP_RELAY` to indicate
whether or not our image should support DHCP relaying. By default as a
heuristic, we use the FLASH_SIZE for the value as only 64MB and bigger
images will need it in practice.

In the future it will be tuned per-platform to be enabled only where
needed and eventually disabled by default.

Change-Id: I08bc07074cada79fcf7b18522d1219ffcc14d1be
Signed-off-by: William A. Kennington III <wak@google.com>