meta-phosphor: initrdscripts: Fix broken PS1 assignments
Commit 51d6af4e64f0 ("meta-phosphor: initrdscripts: fix shellcheck
issues") fixed some shellcheck trailing-space errors the wrong way,
removing the trailing space instead of removing the backslash and
quoting it.  This caused the following lines to be gobbled up into the
variable assignment instead of being executed as their own commands,
which in these particular cases led to kernel panics due to pid 1
exiting (insteading of dropping into an interactive shell as intended).

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>
Change-Id: I8bda0efe825b34abc5219f17fdddc6d7fc57d74f