linux-aspeed: Prepare for CONFIG_EMBEDDED going away
The aspeed configs have EMBEDDED=y, which selects EXPERT. We have a
bunch of options in the config that depend on EXPERT, such as
CONFIG_GPIO_SYSFS.

In v6.6, EMBEDDED is replaced with EXPERT in upstream configs. This is
because the only thing in the kernel that is gated by EMBEDDED is
ARM_PATCH_PHYS_VIRT, which openbmc does not use.

 https://git.kernel.org/torvalds/c/ef815d2cba782e96b9

As they mean the same thing we can make the change to the aspeed kernel
configs now.

Fixes: https://github.com/openbmc/openbmc/issues/3975
Change-Id: Ia2988fd849d26b9540ffb6b064251b8b60e9dca9
Signed-off-by: Joel Stanley <joel@jms.id.au>