meta-facebook: add emmc symlink
We want to transition Bletchley off using /mnt/data as the mount
location for the eMMC device, for consistency with legacy systems
where /mnt/data is for persistent configuration changes in production,
but we first need to create an alternative location for transition
purposes.  Create the /mnt/emmc as a symlink to /mnt/data to facilitate
the transition.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Ie7598a9db43859a8a304c9d20db70639b14908b2