meta-ampere: mtmitchell: handle the fan fault
The phosphor-led-manager will monitors the
`xyz.openbmc_project.State.Decorator.OperationalStatus` interface
of the Inventory D-Bus object, and check whether the Inventory
D-Bus object is associated with the LED group D-Bus object, and
then update the Asserted property of the LedManager.
This adds the fan_fault led group to associations.json file to
monitors the FANs and turns on the fault led when the FANs fails.

1. Unplug the fan and verify that both `Fan Fault Led` and
   `Fault Led` are turned on.
2. Plug the fan and verify that both `Fan Fault Led` and `Fault Led`
   are turned off.

Signed-off-by: Hieu Huynh <hieuh@os.amperecomputing.com>
Change-Id: I2adc0fb8846c4690f8eaac3bb80e6dbf25077be3