meta-facebook: add ability to set remote-addr in ipmb config
Some platforms will need to assign different remote-addr for each
channel. Add a IPMB_REMOTE_ADDR for assigning remote-addr for each
corresponding channel instead of assign fixed remote-addr to all
channels.

If the size of IPMB_REMOTE_ADDR not consistent and less than the size
of IPMB_CHANNEL, the remote-addr will be filled with 64 for the
channels does not assigned.

Test results of tiogapass:
```
{
    "channels": [
        {
            "type": "ipmb",
            "slave-path": "/dev/ipmb-4",
            "bmc-addr": 32,
            "remote-addr": "44",
            "devIndex": 0
        },
        {
            "type": "ipmb",
            "slave-path": "/dev/ipmb-9",
            "bmc-addr": 32,
            "remote-addr": "96",
            "devIndex": 1
        }
    ]
}
```

Signed-off-by: Potin Lai <potin.lai@quantatw.com>
Change-Id: I785201c2c643cf6aa7b83e61df9601ae581f34ad