meta-ibm: Bind MCTP daemon in tandem to PLDM daemon
Applications that operate on BMC typically rely on the standard
systemd recovery mechanism. In this setup, systemd automatically
restarts applications when they encounter issues. However, for
critical daemons like PLDM in IBM systems, the situation is more
nuanced.

PLDM plays a vital role in IBM infrastructures as it facilitates
communication between BMC and the host. Therefore, if the PLDM
daemon experiences a crash or restarts unexpectedly, it's crucial
to notify the remote PLDM endpoints. This notification ensures that
the remote endpoints can detect the disruption and initiate corrective
measures to synchronize critical data, such as PDRs, between the two
endpoints.

To efficiently communicate the status of a remote endpoint when the
PLDM faces an interruption, the MTCP channel serves as a key mechanism.
Specifically, when the mctp-demux daemon undergoes a restart, it
reinitializes the MCTP channel. This action effectively signals the
remote endpoint that an issue has occurred. By doing so, the remote
side can promptly take the necessary corrective actions to maintain
data synchronization and system integrity.

Change-Id: I7d28ef9b3fc8263a100d7ff1f5378fcddcdd3f89
Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>