meta-harma: network: add config 00-bmc-eth0.network
For Meta's machines, we use to use DUID-LL as client identifier for
DHCPv6. We also notice that the IAID is different between AC power on
and reset, which caused the DHCPv6 address is changed due to IAID is
not always consitent.

Add settings to ensure DHCPv6 selects the correct DUID type and forces
IAID to 0 to ensure DHCPv6 address does not change between AC power on
and reset.

```
ClientIdentifier=duid
DUIDType=link-layer
IAID=0
```

Signed-off-by: Potin Lai <potin.lai@quantatw.com>
Change-Id: Ib16959f7f33de6494792eee104fb04ae7ff13eba