meta-facebook: systemd-networkd: wait for one
Many systems have multiple network interfaces, some of which are
USB network interfaces.  We should only wait for a single non-USB
interface to come online in order to consider the `network.target`
as online.  Add a configuration for `systemd-networkd-wait-online`
so that any `ethN` interface coming up is all that is necessary.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I4e5bad889be5a5b9deb537ac5f5df1da33882fad