meta-phosphor: phosphor-state-manager: add host-shutdown to host-stop link
Some systems have separate voltage domains between the chassis and
the host, so they disable the host-shutdown to chassis-poweroff
dependency.  As a side-effect, the host never stops as part of the
host-shutdown.  (The current default dependency is host-shutdown ->
chassis-poweroff -> power-stop -> host-stop ).  Adding an additional
dependency from host-shutdown directly to host-stop, so that the
host is stopped when shutdown.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I7141dbe4ec4590158cc25147a9b387bc562ed957