meta-google: gbmc-bridge: Remove STP support
We no longer have any projects or planned projects that require this
support. We only have single path ethernet as a guarantee now. This
configuration has been causing us some issues on some of our systems and
it will be more stable to just ignore dynamic link configurations and
the software burden of an RSTP stack.

Change-Id: Ia6e113f41d2ad61a79d46a0bf15dd7e4230239db
Signed-off-by: William A. Kennington III <wak@google.com>