meta-phosphor: watchdog: Fix pretimeout time
The name "watchdog pretimeout" in Linux and Systemd is a bit vague.
The watchdog pretimeout is actually the amount of time before the
hardware watchdog timeout that the software watchdog created by the
pretimeout governor will go off.

So for example, the pretimeout of 90 seconds, with a watchdog timeout of
120 seconds, means that the software watchdog will fire off in 30
seconds (120 - 90).

The pretimeout being this early doesn't make much sense, since the
kernel reboots shortly after the software watchdog goes off, so we'll
fix the pretimeout time to be 30 seconds before the hardware watchdog
goes off.

Change-Id: I99a803793fafd8d368387ecac1a451b311b8492b
Signed-off-by: Joe Komlodi <komlodi@google.com>