meta-phosphor: unset VOLATILE_LOG_DIR
This variable, if set, creates a symlink of /var/log to
/var/volatile/log.  In most cases if machines have configured
applications like obmc-console or journald to write to /var/log
they want it to be persistent.  The symlink causes them to be written
to the non-persistent tmpfs space for /var/volatile.

Set all machines to `VOLATILE_LOG_DIR = no` by default.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I0f973e6b8599b32451c83d5161be079e967de711