phosphor state manager: obmcutil package dependency update
obmcutil is currently not enabled for the x86-power-control
based host power managed boxes. This is due to obmcutil
dependency on the chassis package in PSM recipe.
This commits moves the obmcutil package dependency from
chassis -> bmc in PSM recipe to enable tool support in
PSM enabled systems.

Tested: verified build and tested on amd crb.

Change-Id: I19ff062e5a9637cd04111dab5842df39a766b887
Signed-off-by: Jayanth Othayoth <ojayanth@gmail.com>