meta-ibm: introduce new huygens system
The huygens machine is based on the new pstbmc machine family from IBM.
It will be our first redundant BMC system [1].

The initial focus on this machine will be the redundant BMC feature.
The simulation model we use will just be 2 BMCs initially. Due to this
focus and simulation model, this commit strips out all of the host
firmware and host processor related features.

Tested:
- Confirmed in our 2 BMC simulation model that both BMCs come to Ready
  and only the expected vpd services have failed
- Confirmed network into the BMCs worked as expected

[1]: https://gerrit.openbmc.org/c/openbmc/docs/+/70233

Change-Id: Ieb42337980c0ec39d16c492c77810473d5105aa0
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>