poky: reset subtree on master HEAD(a88251b3e7)
As part of the last subtree update, I used master-next as the
subtree basis because there was a fix we needed in order to get QEMU
to pass.  I didn't realize that master-next deviated (because they use
to just have short-term patches in it).  Reset the content to the same
fix but from the master branch.

Change-Id: Ic7d2f0ac42e9da3eb263586b26ba56d8798d5bcf
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>