meta-ibm: bmcweb: enable redfish-query
insecure-enable-redfish-query controls the Redfish expand query
parameter. Named insecure because "is experimental, and has not been
tested against the full limits of user-facing behavior". Downstream we
have done some testing with this enabled and measuring resource usage
are okay with the results.

webui-vue will use expand for sensor page, etc.

Other query parameters such as only are not controlled by this option
and already enabled.

Change-Id: I8403129fabfbe0752c3b0e85bb2e8bd5ebea2aa0
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>