meta-ampere: mtmitchell: update CPU presence object paths
Currently, the gpio-presence is configured to create CPU_1/2 object
paths when sokets 0/1 are presented. But in the SMBIOS service, it is
creating cpu0/1 object paths to indicate sockets 0/1 are presented.
By the missmatch between gpio-presence and SMIBIOS services, adc sensors
that depend on the "CPURequired" property in EM can be missed.

This commit updates the object path of CPU presence to synchronize
gpio-presence and SMBIOS services.

Tested:
    1. PLDM sensors work correctly on both 1P and 2P systems.
    2. ADC sensors are shown correctly.

Change-Id: Idc317b9a6280391aff7eb1aeebc70667acc29b28
Signed-off-by: Thang Tran <thuutran@amperecomputing.com>