Move bmcweb options to meta-ibm
These options don't have anything to do with POWER architecture, they
are ibm options so move them to meta-ibm.

The options are:
    -Dredfish-dbus-log=enabled \
    -Dhttp-body-limit=400 \
and control the body limit, allowing a larger code update image and the
use of phosphor-logging.

This gives the added bonus of all meta-ibm bmcweb options are in one
place and removes these from 2 non-openpower, ibm systems.

Tested: It builds. No other testing.

Change-Id: Ief31f1bfd8a92720b43b4124c6b25c582231a558
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>