meta-facebook: harma: revise PSM behavior
Redfish Command to map PSM behavior:

"ForceOff": Press the power button for over 4 seconds to power off the host.
"PowerCycle": Gracefully power off the host, then power on the host.
"GracefulShutdown": Press the power button for 1 second to gracefully power off the host.
"On": Press the power button for 1 second to power on the host.
"ForceOn": Same as "On".
"GracefulRestart": Same as "PowerCycle".
"ForceRestart": Press the power button for over 4 seconds, then power on the host.

Change-Id: I67b40620107bbf9e90c1c55e37b605ca573f84e4
Signed-off-by: Peter Yin <peter.yin@quantatw.com>