meta-ibm: Add an override for chrony daemon
The huygens machine is based on the new pstbmc machine family from IBM.
It will be our first redundant BMC system [1]. This new system expects
to have chrony daemon to run along with the existing systemd-
timesyncd deamon.
The purpose of installing chrony in rootfs is to run chrony only in
time server mode so that it can allow other BMC to obtain time
from it.

This commit adds an override to the chrony configuration and allows
us to run both the daemons as per our use case.

[1]: https://gerrit.openbmc.org/c/openbmc/docs/+/70233

Tested: Both the daemons run parallely without any conflict.

Change-Id: I9de73ab135977847121daabefaf81265000e371e
Signed-off-by: Pavithra Barithaya <pavithrabarithaya07@gmail.com>