meta-facebook: harma: auto power on
The harma BMC will power on the host
if the CMM performs a host DC cycle or lifts leakage detection.

We can't obtain the correct value from the CPLD
because the MB CPLD has no power at this stage.
Therefore, we need to choose a direct GPIO instead of
the original SGPIO pin.

Change-Id: I0b4c4d3b2950ae489d07972a0bf221074a426d13
Signed-off-by: Peter Yin <peter.yin@quantatw.com>