meta-phosphor: Reuse dreport utility function
install_dreport_user_script function is intended
to be reused in meta-openpower layer.

To enable this, the function is being moved
to a bbclass which would be inherited in the
openpower-debug-collector recipe in the
meta-openpower layer.

dreport.conf file is a variable, hence making it
an argument to the function, so that other layer
can give a custom dreport.conf as input.

Tested:
Verified that plugin scripts are getting installed
from openpower-debug-collector repository and
BMC Dump collection is successful.

Change-Id: I8d13bc7e381cd1b957d5770926fb712165a07185
Signed-off-by: Gopichand Paturi <gopichandpaturi@gmail.com>