meta-ibm: sbp1: virtual_sensor_config: Add SSB power sensors
The SSB provide only current and voltage sensors, but no power
sensors. Assume current and voltages are sampled at the same time
and calculate the output power of each SSB by taking the voltage
and current sensor as input.

Tested: Confirmed that new sensors appear in webui and provide sane
        sensor readings.

Change-Id: Idf187eb01e792ad3716c3ba62fc4792cd770568d
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>