obmc-console: srcrev bump 498a4a81d9..ace05dccbb
Alexander Hansen (1):
      use iniparser dependency for config file parsing

Andrew Jeffery (5):
      subprojects: iniparser: Address meson run_command() warning
      console-server: Avoid SEGV on on failure to load configuration
      config: Allow NULL for internal dict pointer
      console-server: Enable handler iteration with -fsanitize=address
      meson: Exploit `tests` option to disable the test suite

iniparser is a new dependency that replaces the bespoke configuration
parsing implementation. We disable building the tests as that's wasting
cycles for one, but the test suite is also implemented in terms of
iniparser APIs that are not yet available in the version provided by the
bitbake recipe.

Change-Id: I307683a5d6eed1c36e2d0fdd7f99a4e16ce4da3d
Signed-off-by: Andrew Geissler <openbmcbump-github@yahoo.com>
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>