meta-ibm:Use device path for p10bmc hwmon configs
With https://gerrit.openbmc.org/c/openbmc/phosphor-hwmon/+/72996,
phosphor-hwmon can be told to use the device path as the path to the
config file and not the OF_FULLNAME udev env var.

Enable this option for the p10bmc image, and rename all of the config
files in the image to be based on the device path.

This is being done because the Rainier and Everest follow-ons, BlueRidge
and Fuji, have a different name for the max31785 in the device tree,
even though it shares the same fan card.  With this change, the config
files can be shared.

This also deletes the Bonnell config file, because it is at the same
device path as on Rainier but requires different contents.  It will just
be handled with something downstream.

Tested:
Things still show up on D-Bus:
```
$ busctl tree xyz.openbmc_project.Hwmon-75e29b8640dca44a8648ec8aeb240d520bd5eb91b19a8ebe4bbc48ef236d856c.Hwmon1
`- /xyz
  `- /xyz/openbmc_project
    `- /xyz/openbmc_project/sensors
      `- /xyz/openbmc_project/sensors/fan_tach
        |- /xyz/openbmc_project/sensors/fan_tach/fan0_0
        |- /xyz/openbmc_project/sensors/fan_tach/fan0_1
        |- /xyz/openbmc_project/sensors/fan_tach/fan1_0
        |- /xyz/openbmc_project/sensors/fan_tach/fan1_1
        |- /xyz/openbmc_project/sensors/fan_tach/fan2_0
        |- /xyz/openbmc_project/sensors/fan_tach/fan2_1
        |- /xyz/openbmc_project/sensors/fan_tach/fan3_0
        |- /xyz/openbmc_project/sensors/fan_tach/fan3_1
        |- /xyz/openbmc_project/sensors/fan_tach/fan4_0
        |- /xyz/openbmc_project/sensors/fan_tach/fan4_1
        |- /xyz/openbmc_project/sensors/fan_tach/fan5_0
        `- /xyz/openbmc_project/sensors/fan_tach/fan5_1
```

Change-Id: Ie413778b0383cc8115973337dd53d6aa06f45302
Signed-off-by: Matt Spinler <spinler@us.ibm.com>