meta-ibm: Remove local-tty from VUART0 config
local-tty is used as a tty handler where the data read from VUART0 is
copied. The local-tty was set to ttyS0 which is actually hypervisor
console device. That means console server was writing data to hypervisor
console which it read from VUART0. Remove local-tty variable from the
config file same as prior releases.

Change-Id: If5578a47d72acb3a622c934d4d8b0e21a9b4ada8
Signed-off-by: Ninad Palsule <ninad@linux.ibm.com>