phosphor-hwmon: Add use-dev-path packageconfig
With https://gerrit.openbmc.org/c/openbmc/phosphor-hwmon/+/72996,
phosphor-hwmon can be told to use the device path as the path to its
config files and not the OF_FULLNAME udev env var.

Add a 'use-dev-path' packageconfig option to the recipe that bbappend
files in other layers can use to enable this option.

Change-Id: I6009dfbe234fbb8867bbc558b267a5bf45fa878d
Signed-off-by: Matt Spinler <spinler@us.ibm.com>