Add Google Authenticator PAM module for 2FA
Google authenticator PAM module enables two-factor authentication for
logging in to servers. It provides both HMAC-based and Time-based OTPs
as the second factor on top of usual password authentication.

This commit adds a new distro feature for google authenticator and can
be enabled or disabled for any machine based on the requirement.

Tested By:
* Built openbmc for p10bmc
* Verified that the google-authenticator was pulled and the binary is
  installed

Change-Id: I586715a924eddd90c9cb6f967db35824767330ca
Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>