meta-google: gbmc-bridge: Send hostname in RAs
This adds support for a process that monitors hostname changes and adds
the current hostnmae as a "DNS Search List" field entry in the RA
messages from a node. This makes it possible to discover the hostnames
of nodes within the machine simply by looking at RA messages.

Tested: Ran on a machine and verified with `rdisc6 gbmcbr` that the
hostname was being included in RA messages and updated after
`hostnamectl set-hostname <new name>`

Change-Id: Iac8ee6300384191f4fb8ec3bf252afe84ef814c0
Signed-off-by: William A. Kennington III <wak@google.com>