meta-google: gbmc-bridge: Refactor ip from bridge script
This refactors the RA querying and parsing logic so that other script
can leverage the functionality if they want to discover hostname and IP
information from RAs. There are no functional changes in this commit.

Tested: Verified that the script still determines the NCSI hostname and
IP correctly on machines using the script.

Change-Id: I751048440f2e9eab1628600a9bdc3701f07adf76
Signed-off-by: William A. Kennington III <wak@google.com>