meta-google: gbmc-bridge: Add optional fixed IP offset
Some of our machines do not support receiving DHCP packets to determine
their addresses, and need a workaround to piggyback off DHCPd nodes in
the system.

This change allows a machine to set a fixed numeric offset address,
where it can read the correct BMC prefix from RAs of other nodes and
then generate a fixed address for itself.

Tested: Ran on a 2 node BMC machine with one using this fixed scheme.
Verified that it correctly found an address and hostname.

Change-Id: Ibd79f800a582db44cf2ad469dbe0a75f13dcaeed
Signed-off-by: William A. Kennington III <wak@google.com>