meta-facebook: yosemite4: adjust init script for yosemite4n
Adjusting the initialization script for the yosemite4n platform
to support Boot A/B Recovery for the Nuvoton platform.

The key change involves using the INTCR2.WDC register to identify the
flash source, making the script compatible with both Aspeed and Nuvoton.

Tested:
When booting from the primary flash, the slot value is 0:
root@bmc:~# cat /run/media/slot
0

When the recovery mechanism is triggered,
booting from the alternate flash, the slot value is 1:
root@bmc:~# cat /run/media/slot
1

Signed-off-by: Tim Lee <timlee660101@gmail.com>
Change-Id: Ibd3c53094d6577e73da545f257ae47db3df6cb69