meta-ibm: system1: Add bios version package
This commit includes changes to update the bios version flashed in x86
PCH SPI flash. The version is updated in the dbus as well and cached for
later use to avoid reading SPI flash subsequently. This service is
started after "xyz.openbmc_project.Software.Version.service", typically
during boot or when flashing a host firmware via BMC to update version
information.

Tested By:
Built openbmc, and verified that the service is installed on the bmc.
The host firmware version is updated on the dbus [1] and updated in the
`/var/cache/bios_version` file, when PCH is in standby.

[1] busctl get-property xyz.openbmc_project.Software.BMC.Updater /xyz/openbmc_project/software/bios_active xyz.openbmc_project.Software.Version Version

Change-Id: Ic7b4bb1250b8f1b8fc62fc7fd46654375937a459
Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>