meta-facebook: harma: remove flow control
We removed the flow control flag and used DMA as an alternative
because flow control has a limitation requiring a handshake redo
with the BMC. This issue prevents us from connecting to the host
after a BMC reboot.

The root cause of the previous kernel panic issue was that
the Aspeed register was not being cleared when we enable DMA
mode. Therefore, we need to perform an AC cycle whenever we enable
or disable the DMA version change.

Change-Id: I25a595c2625a63c45f453cb8546a006f72b1871b
Signed-off-by: Peter Yin <peter.yin@quantatw.com>